services/financial-agents/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.3

# TradingAgents (TauricResearch)
tradingagents-framework==0.1.4

# FinRobot (AI4Finance-Foundation) — optional enrichment layer
# Comment out if you don't want the longer install time
finrobot==0.4.2
pyautogen==0.3.1
